% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_unlinked_tree.R
\name{add_hidden_nodes}
\alias{add_hidden_nodes}
\title{calculate the expected number of hidden nodes, and draw from Poisson
distribution to obtain the realized number of hidden nodes}
\usage{
add_hidden_nodes(phy, lambda, mu)
}
\arguments{
\item{phy}{phylogeny}

\item{lambda}{birth rate}

\item{mu}{death rate}
}
\description{
calculate the expected number of hidden nodes, and draw from Poisson
distribution to obtain the realized number of hidden nodes
}
