% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{function to calculate the number of substitutions along a tree}
\usage{
calc_dist(alignment_phydat, root_sequence)
}
\arguments{
\item{alignment_phydat}{alignment object}

\item{root_sequence}{root sequence}
}
\description{
function to calculate the number of substitutions along a tree
}
