% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_expected_fraction}
\alias{calc_expected_fraction}
\title{calculate the expected fraction of substitutions at the nodes, relative to the fraction
at the branches}
\usage{
calc_expected_fraction(
  phy = NULL,
  num_tips = NULL,
  node_time,
  model = "unlinked",
  lambda = NULL,
  mu = NULL
)
}
\arguments{
\item{phy}{phylogenetic tree (optional)}

\item{num_tips}{number of tips (optional)}

\item{node_time}{time spent at the node}

\item{model}{node substitution model}

\item{lambda}{the birth rate to generate the phylogenetic tree}

\item{mu}{the death rate to generate the phylogenetic tree}
}
\value{
expected fraction
}
\description{
calculates the relative contribution of substitutions at the nodes
}
