% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_expected_hidden_nodes_per_branch}
\alias{calc_expected_hidden_nodes_per_branch}
\title{calculate the number of expected hidden nodes along a branch}
\usage{
calc_expected_hidden_nodes_per_branch(bl, lambda, mu)
}
\arguments{
\item{bl}{branch length}

\item{lambda}{birth rate}

\item{mu}{death rate}
}
\value{
expected number of hidden nodes
}
\description{
calculate the number of expected hidden nodes by
integrating over equation 1 in Maceau et al. 2019
}
