% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_fraction}
\alias{calc_fraction}
\title{Calculate the expected fraction of substitutions at the nodes,
relative to the fraction at the branches}
\usage{
calc_fraction(phy = NULL, node_time = 0, model = "unlinked")
}
\arguments{
\item{phy}{phylogenetic tree (optional)}

\item{node_time}{time spent at the node}

\item{model}{node substitution model}
}
\value{
expected fraction
}
\description{
calculates the relative contribution of substitutions at
the nodes
}
