% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_rel_contr_node_time}
\alias{calc_rel_contr_node_time}
\title{calculate the fraction of substitutions at the nodes, relative to the fraction
at the branches}
\usage{
calc_rel_contr_node_time(phy, node_time, lambda = NULL, mu = NULL)
}
\arguments{
\item{phy}{phylogenetic tree}

\item{node_time}{time spent at the node}

\item{lambda}{the birth rate to generate the phylogenetic tree}

\item{mu}{the death rate to generate the phylogenetic tree}
}
\value{
required fraction
}
\description{
calculates the relative contribution of substitutions at the nodes
}
