% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_required_node_time}
\alias{calc_required_node_time}
\title{calculate the required node time to obtain a desired fraction of substitions at the node}
\usage{
calc_required_node_time(phy = NULL, num_tips = NULL, s, model = "unlinked")
}
\arguments{
\item{phy}{phylogenetic tree (optional)}

\item{num_tips}{number of tips (optional)}

\item{s}{desired fraction}

\item{model}{node substitution model}
}
\value{
expected fraction
}
\description{
calculates the required node time to obtain a desired fraction of susbstitutions at the node
}
