% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_statistics.R
\name{calc_sum_stats}
\alias{calc_sum_stats}
\title{calculate summary statistics compared with a reference tree}
\usage{
calc_sum_stats(trees, true_tree, verbose = FALSE)
}
\arguments{
\item{trees}{a phyloList object containing multiple trees}

\item{true_tree}{a phylo object containing the reference tree}
}
\value{
list with two tibbles 1) containing the summary statistics of all trees and 2) containing the difference with the true tree
}
\description{
calculate summary statistics compared with a reference tree
}
