% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_spent_at_node.R
\name{calc_time_spent_at_node}
\alias{calc_time_spent_at_node}
\title{calculate the expected time spent on the node
in the nodes}
\usage{
calc_time_spent_at_node(phy, node_time = NULL, lambda = NULL,
  mu = NULL, is_birth_death = FALSE, model = "parent")
}
\arguments{
\item{phy}{phylogenetic tree}

\item{node_time}{amount of time spent at the node}

\item{lambda}{the birth rate to generate the phylogenetic tree}

\item{mu}{the death rate to generate the phylogenetic tree}

\item{is_birth_death}{if false, a yule tree is}

\item{model}{the used node substitution model}
}
\value{
expected fraction of time spent on the nodes
}
\description{
calculates the required node time
}
\details{
if no birth or death rates are provided, birth and death rates are
estimated by fitting a birth-death model to the tree (or a yule tree)
}
