% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_hidden.R
\name{count_hidden}
\alias{count_hidden}
\title{function to calculate the number of hidden speication events}
\usage{
count_hidden(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
number of hidden speciation eventsß
}
\description{
function to calculate the number of hidden speication events
}
