% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_from_brts.R
\name{create_balanced_tree}
\alias{create_balanced_tree}
\title{create a balanced tree out of branching times}
\usage{
create_balanced_tree(brts)
}
\arguments{
\item{brts}{vector of branching times}
}
\value{
phylo phylo object
}
\description{
create a balanced tree out of branching times
}
