% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equal_alignment.R
\name{create_equal_alignment}
\alias{create_equal_alignment}
\title{function create an alignment with highly similar information content}
\usage{
create_equal_alignment(
  input_tree,
  fraction,
  sub_rate,
  focal_alignment,
  root_sequence,
  alt_model,
  verbose = FALSE
)
}
\arguments{
\item{input_tree}{phylogeny for which to generate alignment}

\item{fraction}{fraction of time spent on the nodes}

\item{sub_rate}{substitution rate used in the original phylogeny}

\item{focal_alignment}{alignment to match information content with}

\item{root_sequence}{root sequence}

\item{alt_model}{alternative substitution model}

\item{verbose}{provide intermediate output}
}
\value{
list with alignment and inferred rate
}
\description{
function create an alignment with highly similar information content
}
