% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equal_alignment_explicit.R
\name{create_equal_alignment_explicit}
\alias{create_equal_alignment_explicit}
\title{function create an alignment with identical information content, using the
explicit method to simulate substitutions}
\usage{
create_equal_alignment_explicit(
  input_tree,
  sub_rate,
  alignment_result,
  verbose = FALSE
)
}
\arguments{
\item{input_tree}{phylogeny for which to generate alignment}

\item{sub_rate}{substitution rate used in the original phylogeny}

\item{alignment_result}{result of sim_normal, sim_linked or sim_unlinked}

\item{verbose}{provide intermediate output}
}
\value{
list with four properties: 1) alignment: the alignment itself,
2) adjusted rate: the substitution rate used to obtain identical information
content 3) total_accumulated_substitutions: the total number of
substitutions accumulated. 4) total_node_substitutions: total number of
substitutions accumulated on the nodes 5) total_branch_substitutions: total
number of substitutions accumulated on the branches.
}
\description{
function create an alignment with identical information content, using the
explicit method to simulate substitutions
}
