% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_unlinked_tree.R
\name{create_equal_alignment_tree}
\alias{create_equal_alignment_tree}
\title{simulate a sequence assuming node substitutions are shared among the offspring
optimize to obtain an equal amount of substitutions as a given alignment}
\usage{
create_equal_alignment_tree(
  input_tree,
  focal_alignment = NULL,
  Q1 = NULL,
  Q2 = NULL,
  rate1 = 1,
  rate2 = 1,
  l = 1000,
  bf = NULL,
  rootseq = NULL,
  node_time = 0.001,
  lambda = NULL,
  mu = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input_tree}{tree for which to simulate sequences}

\item{focal_alignment}{alignment to match information content with}

\item{Q1}{substitution matrix along the branches, default = JC}

\item{Q2}{substitution matrix on the nodes, default = JC}

\item{rate1}{mutation rate along the branch, default = 1}

\item{rate2}{mutation rate on the node, default = 1}

\item{l}{number of base pairs to simulate}

\item{bf}{base frequencies, default = c(0.25, 0.25, 0.25, 0.25)}

\item{rootseq}{sequence at the root, simulated by default}

\item{node_time}{amount of time spent at the nodes}

\item{lambda}{birth rate}

\item{mu}{death rate}

\item{verbose}{should intermediate output be displayed?}
}
\value{
phyDat object
}
\description{
simulate a sequence assuming node substitutions are shared among the offspring
optimize to obtain an equal amount of substitutions as a given alignment
}
