% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_from_brts.R
\name{create_unbalanced_tree}
\alias{create_unbalanced_tree}
\title{create an unbalanced tree out of branching times}
\usage{
create_unbalanced_tree(brts)
}
\arguments{
\item{brts}{vector of branching times}
}
\value{
phylo phylo object
}
\description{
create an unbalanced tree out of branching times
}
