% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_marginal_models.R
\name{estimate_marginal_models}
\alias{estimate_marginal_models}
\title{infer the time calibrated phylogeny associated with the}
\usage{
estimate_marginal_models(fasta_filename, rng_seed = 42, verbose = FALSE)
}
\arguments{
\item{fasta_filename}{file name of fasta file holding alignment for which the
marginal likelihood is to be estimated}

\item{rng_seed}{seed of pseudo-random number generator}

\item{verbose}{boolean indicating if verbose intermediate output is to be
generated}
}
\value{
data frame
}
\description{
infer the time calibrated phylogeny associated with the
}
