% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_normal_branching_rate.R
\name{find_normal_branching_rate}
\alias{find_normal_branching_rate}
\title{find a matching substitution rate}
\usage{
find_normal_branching_rate(phy, alignment, sim_regular, replicates = 1,
  max_rate = 1)
}
\arguments{
\item{phy}{tree for which to simulate sequences}

\item{alignment}{alignment of alternative model}

\item{sim_regular}{function to simulate an alignment under a normal model, taking two arguments: phy and rate}

\item{replicates}{number of replicate runs for which to estimate}

\item{max_rate}{maximum rate to be considered in optimization}
}
\value{
rate
}
\description{
find a matching substitution rate
}
