% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_tree}
\alias{get_tree}
\title{generate a tree conditional on n and t}
\usage{
get_tree(b, m, focal_num_tips, focal_crown_age)
}
\arguments{
\item{b}{birth rate (speciation rate)}

\item{m}{death rate (extinction rate)}

\item{focal_num_tips}{total number of tips to be conditioned on}

\item{focal_crown_age}{crown age to be conditioned on}
}
\value{
phylo tree
}
\description{
generates a tree, conditional on number of tips and crown age
}
