% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getP_thijs}
\alias{getP_thijs}
\title{testing function for calculation of matrix}
\usage{
getP_thijs(el, eig = phangorn::edQt(), g = 1)
}
\arguments{
\item{el}{branch length}

\item{eig}{list with eigenvectors/values}

\item{g}{rate}
}
\description{
testing function for calculation of matrix
}
