% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_phylogeny.R
\name{infer_phylogeny}
\alias{infer_phylogeny}
\title{infer the time calibrated phylogeny associated with the}
\usage{
infer_phylogeny(alignment, treatment_name, burnin, chain_length)
}
\arguments{
\item{alignment}{Phydat object containing the focal alignment}

\item{treatment_name}{string to be appended to BEAST files}

\item{burnin}{burnin of posterior distribution}

\item{chain_length}{length of the MCMC chain}
}
\value{
list with all trees, and the consensus tree
}
\description{
infer the time calibrated phylogeny associated with the
}
