% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_phylogeny_2.R
\name{infer_phylogeny_2}
\alias{infer_phylogeny_2}
\title{infer the time calibrated phylogeny associated with the}
\usage{
infer_phylogeny_2(alignment, treatment_name,
  inference_model = beautier::create_inference_model(site_model =
  beautier::create_jc69_site_model(), clock_model =
  beautier::create_strict_clock_model(), tree_prior =
  beautier::create_yule_tree_prior(), mcmc =
  beautier::create_mcmc(chain_length = 1e+06, store_every = 5000)),
  mcmc_seed = NULL, burnin, working_dir = NULL)
}
\arguments{
\item{alignment}{Phydat object containing the focal alignment}

\item{treatment_name}{string to be appended to BEAST files}

\item{inference_model}{inference model used in BEAST analysis,
as generated by the function beautier::create_inference_model}

\item{mcmc_seed}{seed of the mcmc chain, default is the system time}

\item{burnin}{burnin of posterior distribution}

\item{working_dir}{beast2 working dir}
}
\value{
list with all trees, and the consensus tree
}
\description{
infer the time calibrated phylogeny associated with the
}
