\name{nodeSub-package}
\alias{nodeSub-package}
\alias{nodeSub}
\docType{package}
\title{
	Package providing functions to simulate sequences under different DNA evolution models
}
\description{
	This package provides functions to simulate sequences under different DNA evolution models

}
\details{
\tabular{ll}{
Package: \tab nodeSub\cr
Type: \tab Package\cr
Version: \tab 0.01 \cr
Date: \tab 2019-11-04\cr
License: GPL 2.0
}

}

\author{
Thijs Janzen

Maintainer: Thijs Janzen <t.janzen@rug.nl>
}
