% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_hidden.R
\name{reduce_tree}
\alias{reduce_tree}
\title{Function to remove speciation events occuring after an extinction event.
Extinct species are pruned randomly, such that only a single extinct species
per branching event (if any extinct species) remains.}
\usage{
reduce_tree(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
pruned tree
}
\description{
Function to remove speciation events occuring after an extinction event.
Extinct species are pruned randomly, such that only a single extinct species
per branching event (if any extinct species) remains.
}
