% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dual_independent.R
\name{sim_dual_independent}
\alias{sim_dual_independent}
\title{simulate a sequence assuming node substitutions are shared among the offspring}
\usage{
sim_dual_independent(
  phy,
  Q1 = NULL,
  Q2 = NULL,
  rate1 = 1,
  rate2 = 1,
  l = 1000,
  bf = NULL,
  rootseq = NULL,
  node_time = 0.001
)
}
\arguments{
\item{phy}{tree for which to simulate sequences}

\item{Q1}{substitution matrix along the branches, default = JC}

\item{Q2}{substitution matrix on the nodes, default = JC}

\item{rate1}{mutation rate along the branch, default = 1}

\item{rate2}{mutation rate on the node, default = 1}

\item{l}{number of base pairs to simulate}

\item{bf}{base frequencies, default = c(0.25, 0.25, 0.25, 0.25)}

\item{rootseq}{sequence at the root, simulated by default}

\item{node_time}{amount of time spent at the nodes}
}
\value{
phyDat object
}
\description{
simulate a sequence assuming node substitutions are shared among the offspring
}
