% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dual_linked.R
\name{sim_dual_subs_linked}
\alias{sim_dual_subs_linked}
\title{simulate a sequence assuming conditional substitutions on the node}
\usage{
sim_dual_subs_linked(phy, Q = NULL, rate = 1, mu = 1e-09, l = 1000,
  bf = NULL, rootseq = NULL)
}
\arguments{
\item{phy}{tree for which to simulate sequences}

\item{Q}{substitution matrix along the branches, default = JC}

\item{rate}{mutation rate , default = 1}

\item{mu}{mutation rate on the node, default = 1e-9}

\item{l}{number of base pairs to simulate}

\item{bf}{base frequencies, default = c(0.25, 0.25, 0.25, 0.25)}

\item{rootseq}{sequence at the root, simulated by default}
}
\value{
phyDat object
}
\description{
simulate a sequence assuming conditional substitutions on the node
}
