% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_normal.R
\name{sim_normal}
\alias{sim_normal}
\title{Simulate sequences.}
\usage{
sim_normal(x, l = 1000, Q = NULL, bf = NULL, rootseq = NULL,
  rate = 1)
}
\arguments{
\item{x}{a phylogenetic tree \code{tree}, i.e. an object of class
\code{phylo} or and object of class \code{pml}.}

\item{l}{length of the sequence to simulate.}

\item{Q}{the rate matrix.}

\item{bf}{base frequencies.}

\item{rootseq}{a vector of length l containing the root sequence, other root
sequence is randomly generated.}

\item{rate}{mutation rate or scaler for the edge length, a numerical value
greater than zero.}
}
\value{
\code{simSeq} returns an object of class phyDat.
}
\description{
Simulate sequences for a given evolutionary tree.
}
\details{
\code{simSeq} is now a generic function to simulate sequence alignments to
along a phylogeny. It
is quite flexible and allows to generate DNA, RNA, amino acids, codon  or
binary sequences.  It is possible to give a \code{pml} object as input simSeq
return a \code{phyDat} from these model.  There is also a more low level
version, which lacks rate variation, but one can combine different
alignments having their own rate (see example). The rate parameter acts like
a scaler for the edge lengths.

can be supplied.
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
