% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_along_branch.R
\name{simulate_along_branch}
\alias{simulate_along_branch}
\title{simulate substitutions along an arbitrary, imaginary, branch.}
\usage{
simulate_along_branch(
  t0,
  t1,
  lambda,
  mu,
  node_time,
  Q = NULL,
  bf = NULL,
  l,
  mut_rate,
  node_rate
)
}
\arguments{
\item{t0}{starting time, assuming t1 > t0 > 0}

\item{t1}{end time, assuming t1 > t0 > 0}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{node_time}{time spent on the nodes}

\item{Q}{substitution matrix, default = JC}

\item{bf}{base frequencies, default = c(0.25, 0.25, 0.25, 0.25)}

\item{l}{number of base pairs to simulate}

\item{mut_rate}{mutation rate along the branch}

\item{node_rate}{mutation rate at the node}
}
\value{
list with number of substitutions
}
\description{
simulate substitutions along an arbitrary, imaginary, branch.
}
