% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{slow_matrix}
\alias{slow_matrix}
\title{this function calculates the p matrix within R
this is slower than the C++ implementation
but provides a way to debug}
\usage{
slow_matrix(eig, branch_length, rate)
}
\description{
this function calculates the p matrix within R
this is slower than the C++ implementation
but provides a way to debug
}
