% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_required_node_time.R
\name{calc_required_node_time}
\alias{calc_required_node_time}
\title{calculate the required node time to obtain a % of all substitutions
in the nodes}
\usage{
calc_required_node_time(phy, fraction, lambda = NULL, mu = NULL,
  is_birth_death = FALSE, model = "parent")
}
\arguments{
\item{phy}{phylogenetic tree}

\item{fraction}{the targeted fraction}

\item{is_birth_death}{if false, a yule tree is assumed}

\item{model}{the used node substitution model}

\item{birth_rate}{the birth rate to generate the phylogenetic tree}

\item{death_Rate}{the death rate to generate the phylogenetic tree}
}
\value{
required fraction
}
\description{
calculates the required node time
}
\details{
if no birth or death rates are provided, birth and death rates are
estimated by fitting a birth-death model to the tree (or a yule tree)
}
