% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_contr_node_time.R
\name{calc_required_node_time}
\alias{calc_required_node_time}
\title{Calculate the required node time to obtain a desired fraction of
substitutions at the node}
\usage{
calc_required_node_time(phy = NULL, s = 0.5, model = "unlinked")
}
\arguments{
\item{phy}{phylogenetic tree}

\item{s}{desired fraction}

\item{model}{node substitution model, either "linked" or "unlinked".}
}
\value{
expected fraction
}
\description{
calculates the required node time to obtain a desired fraction
of substitutions at the node
}
