% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_hidden.R
\name{count_hidden}
\alias{count_hidden}
\title{Function to calculate the number of hidden speciation events, e.g. speciation
events that have lead to an extinct species. Thus, these hidden speciation
events can only be detected in complete trees (as opposed to reconstructed
trees).}
\usage{
count_hidden(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
number of hidden speciation events
}
\description{
Function to calculate the number of hidden speciation events, e.g. speciation
events that have lead to an extinct species. Thus, these hidden speciation
events can only be detected in complete trees (as opposed to reconstructed
trees).
}
