% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equal_alignment.R
\name{create_equal_alignment}
\alias{create_equal_alignment}
\title{function create an alignment with highly similar information content}
\usage{
create_equal_alignment(input_tree, focal_alignment, root_sequence,
  alt_model)
}
\arguments{
\item{input_tree}{phylogeny for which to generate alignment}

\item{focal_alignment}{alignment to match information content with}

\item{root_sequence}{root sequence}

\item{alt_model}{alternative substitution model}
}
\value{
list with alignment and inferred rate
}
\description{
function create an alignment with highly similar information content
}
