% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equal_alignment.R
\name{create_equal_alignment}
\alias{create_equal_alignment}
\title{function create an alignment with highly similar information content}
\usage{
create_equal_alignment(input_tree, sub_rate, alignment_result, verbose = FALSE)
}
\arguments{
\item{input_tree}{phylogeny for which to generate alignment}

\item{sub_rate}{substitution rate used in the original phylogeny}

\item{alignment_result}{result of sim_normal, sim_linked or sim_unlinked}

\item{verbose}{provide intermediate output}
}
\value{
list with alignment and inferred rate
}
\description{
function create an alignment with highly similar information content
}
