% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_normal_branching_rate.R
\name{find_normal_branching_rate}
\alias{find_normal_branching_rate}
\title{find a matching substitution rate}
\usage{
find_normal_branching_rate(phy, alignment, sim_regular, replicates = 1,
  max_rate = 1)
}
\arguments{
\item{phy}{tree for which to simulate sequences}

\item{sim_regular}{function to simulate an alignment under a normal model, taking two arguments: phy and rate}

\item{sim_nodesub}{function to simulate an alignment under a node substitution model}

\item{num_sim_nodesub}{how many alignments to compare with should be generated?}

\item{num_sim_regular}{how many alignments to compare with a single nodesub alignment should be generated?}
}
\value{
rate
}
\description{
find a matching substitution rate
}
