\name{nodeSub-package}
\alias{nodeSub-package}
\alias{nodeSub}
\docType{package}
\title{
	Package providing functions to simulate sequences under different DNA evolution models
}
\description{
	This package provides functions to simulate sequences under different DNA evolution models

}

\author{
Thijs Janzen
Maintainer: Thijs Janzen <t.janzen@rug.nl>
}

\references{
Janzen, T., Bokma, F. and Etienne, R.S. Nucleotide substitutions during speciation may explain substitution rate variation (2021) bioRxiv 2020.08.19.256891
}
