% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{slow_matrix}
\alias{slow_matrix}
\title{this function calculates the p matrix within R
this is slower than the C++ implementation in \code{get_p_matrix}
but provides a way to debug and verify}
\usage{
slow_matrix(eig, branch_length, rate)
}
\arguments{
\item{eig}{eigen object}

\item{branch_length}{branch length}

\item{rate}{substitution rate}
}
\value{
p matrix
}
\description{
this function calculates the p matrix within R
this is slower than the C++ implementation in \code{get_p_matrix}
but provides a way to debug and verify
}
